MAINT: refactor sklearn dependencies to separate module
